Show layout/parse errors above the graph instead of in the Layout tab
The spytial-core error modal was mounted inside GraphLayoutDrawer, so it
only rendered when the Layout drawer tab was active — errors raised by
SpyTialGraph were invisible on every other tab (Time, Projections,
Explorer, Synthesis, Log). It also re-created its React root on each tab
switch.

Mount the error modal once in GraphView, directly above the graph, where
it is visible regardless of the active drawer tab. The modal collapses to
zero height when there is no error and reflows the graph down when one is
present. Extract ensureBootstrapLoaded() so the error card's Bootstrap
styles are available from the graph view (previously only the Layout
drawer loaded them).
