+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "sterling-layout",
   "description": "Cope and Drag (CnD): a fork of the Sterling visualizer that encodes spatial layout.",
-  "version": "4.0.5",
+  "version": "4.0.6",
   "main": "index.js",
   "author": "Siddhartha Prasad",
   "license": "MIT",
